Regulated file names.  Deleted unnecessary files.
Edits on 2017-09-14.
This is still a 'reconstruction' of the codes.
(1) I removed some of the files that I think are redundant.
(2) I regulated the file names, so they have the same name as the codes on my computer.  But I didn't change the contents of the codes comparing to the last commit.
(3) Not all codes on my computer are copied here.  But I will add them later.

So, this commit represents an intermediate state of editing.